svn merge -r15417:15598  https://eris.llnl.gov/svn/slurm/branches/priority

+++ trunk/src/common/jobacct_common.h
@@ -133,7 +133,7 @@ typedef struct {
 	uint32_t lft;
 	char	*partition;
 	char	*nodes;
-	int32_t priority;
+	double priority;
 	uint16_t qos;
 	uint32_t req_cpus;
 	uint32_t requid;
